use box2d<float> in relevant in spatial_index's [experimental]

+++ benchmark/test_quad_tree.cpp
@@ -20,7 +20,7 @@ public:
         std::random_device rd;
         std::default_random_engine engine(rd());
         std::uniform_int_distribution<int> uniform_dist(0, 2048);
-        quad_tree_type tree(mapnik::box2d<double>(0,0,2048,2048));
+        quad_tree_type tree(mapnik::box2d<float>(0,0,2048,2048));
         //populate
         for (size_t i = 0; i < iterations_; ++i)
         {
@@ -28,7 +28,7 @@ public:
             int cy = uniform_dist(engine);
             int sx = 0.2 * uniform_dist(engine);
             int sy = 0.2 * uniform_dist(engine);
-            mapnik::box2d<double> box(cx - sx,cy - sy, cx + sx, cy + sy);
+            mapnik::box2d<float> box(cx - sx,cy - sy, cx + sx, cy + sy);
             tree.insert(i, box);
         }
         // bounding box query
@@ -39,7 +39,7 @@ public:
             int cy = uniform_dist(engine);
             int sx = 0.4 * uniform_dist(engine);
             int sy = 0.4 * uniform_dist(engine);
-            mapnik::box2d<double> box(cx - sx,cy - sy, cx + sx, cy + sy);
+            mapnik::box2d<float> box(cx - sx,cy - sy, cx + sx, cy + sy);
             auto itr = tree.query_in_box(box);
             auto end = tree.query_end();
             for ( ;itr != end; ++itr)
